requirements.txt
irc
python-daemon >= 1.6
requests
icalendar >= 5.0.0
